feat: Transaction edits

+++ src/main/java/com/banking/service/UserService.java
@@ -34,7 +34,7 @@ public class UserService
         return userRepository.findByUsername(username);
     }
 
-    public int getIdByUsername(String username)
+    public Long getIdByUsername(String username)
     {
         User user = userRepository.findByUsername(username);
         return user != null ? user.getId() :  -1;
